Fix logic causing E0005 and E0381

+++ crates/conductor_lib/src/conductor/admin.rs
@@ -247,9 +247,11 @@ impl ConductorAdmin for Conductor {
 
         if let Some(instance) = self.instances.remove(id) {
             if clean {
-                let Some(instance_config) = self.config.instance_by_id(id);
-                let Some(storage_path) = instance_config.storage.get_path();
-                remove_dir_all(storage_path)?;
+                if let Some(instance_config) = self.config.instance_by_id(id) {
+                    if let Some(storage_path) = instance_config.storage.get_path() {
+                        remove_dir_all(storage_path)?;
+                    }
+                }
             }
             instance.write().unwrap().kill();
         }
